+++ app/dependencies/order_dependency.py
@@ -11,7 +11,7 @@ from entities.order.choices import StatusOrder
 from models.database import get_session
 from entities.order.models_order import OrderInSchema
 from entities.order.schemas.schema_order import OrderPaymentSchema
-from services.order_service import OrderService
+from entities.order.services_order import OrderService
 from entities.users.services.client_service import ClientService
 from entities.offer.services_offer import OfferLinkService
 from entities.service_name.services_service_name import ServiceNameService
